requirements.txt

ipdb
pytest
matplotlib
numpy==1.23.5
pandas==2.0.2
python-dotenv

# for api (fastAPI)
fastapi
starlette
uvicorn

# for app (streamlit)
streamlit
pydeck
Pillow

# cloud 
google-cloud-storage


# Tensorflow for modelling

# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.12.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.12.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.12.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips


